# backend/lib/internal/server/handlers/openapi.yaml
openapi: 3.0.2
info:
  title: Save to Ink API
  description: REST API for article management and delivery service
  version: 0.33.9-dev+19032230.b7e98cd
servers:
  - url: /v1
    description: API base path
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Article:
      type: object
      required:
        - account
        - id
        - url
        - createdAt
      properties:
        account:
          type: string
          description: User account ID
        id:
          type: string
          description: Article ID
        url:
          type: string
          description: Article URL
        createdAt:
          type: string
          format: date-time
          description: Article creation timestamp
        title:
          type: string
          description: Article title
        content:
          type: string
          description: Article content
        author:
          type: string
          description: Article author
        siteName:
          type: string
          description: Site name
        sourceDomain:
          type: string
          description: Source domain
        excerpt:
          type: string
          description: Article excerpt
        imageUrl:
          type: string
          description: Image URL
        contentType:
          type: string
          description: Content type
        language:
          type: string
          description: Article language
        error:
          type: string
          description: Error message if processing failed
        wordCount:
          type: integer
          description: Word count
        readingTimeMinutes:
          type: integer
          description: Estimated reading time in minutes
        publishedAt:
          type: string
          format: date-time
          description: Original publication date
        favorite:
          type: boolean
          description: Whether article is favorited
    ArticleRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: URL of the article to save (http or https only, private IPs blocked)
          maxLength: 2000
        send_on_complete:
          type: boolean
          description: Whether to send the article to the user's device after processing is complete
          default: false
    ArticleResponse:
      type: object
      required:
        - id
        - title
        - url
      properties:
        id:
          type: string
          description: Article ID
        title:
          type: string
          description: Article title
        url:
          type: string
          description: Article URL
    ListArticlesResponse:
      type: object
      required:
        - articles
        - page
        - page_size
        - total
        - has_more
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
          description: List of articles
        page:
          type: integer
          description: Current page number
        page_size:
          type: integer
          description: Number of articles per page
        total:
          type: integer
          description: Total number of articles
        has_more:
          type: boolean
          description: Whether there are more pages
    DeleteArticleResponse:
      type: object
      required:
        - deleted
      properties:
        deleted:
          type: integer
          description: Number of articles deleted
    FavoriteResponse:
      type: object
      required:
        - favorite
      properties:
        favorite:
          type: boolean
          description: New favorite status
...
    UserProfile:
      type: object
      required:
        - account
        - email
        - device_email
        - auto_send
      properties:
        account:
          type: string
          description: User account ID
        email:
          type: string
          description: User email address
        device_email:
          type: string
          description: User e-reader device email
        auto_send:
          type: boolean
          description: Whether to auto-send articles
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "invalid URL: URL points to private/internal network"
...
paths:
...
  /articles:
    get:
      summary: List articles
      description: Get a paginated list of articles for the authenticated user
      operationId: listArticles
      tags:
        - Articles
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number, default is 1
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Number of articles per page, default is 20, max is 100
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: favorite
          in: query
          description: Filter by favorite status
          schema:
            type: boolean
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListArticlesResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      summary: Create article
      description: Save a new article from a URL. If send_on_complete is true, checks quota and device email status before creating.
      operationId: createArticle
      tags:
        - Articles
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleRequest"
      responses:
        "201":
          description: Article created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleResponse"
        "400":
          description: Bad request (invalid URL, email backend not configured, or device email is bouncing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict (device email is bouncing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too Many Requests (free tier quota exceeded)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/{id}:
    get:
      summary: Get article
      description: Get a single article by ID
      operationId: getArticle
      tags:
        - Articles
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Article ID
          schema:
            type: string
      responses:
        "200":
          description: Article details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Article not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete article
      description: Delete a single article by ID
      operationId: deleteArticle
      tags:
        - Articles
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Article ID
          schema:
            type: string
      responses:
        "200":
          description: Article deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteArticleResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/{id}/favorite:
    put:
      summary: Toggle favorite
      description: Toggle the favorite status of an article
      operationId: toggleFavorite
      tags:
        - Articles
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Article ID
          schema:
            type: string
      responses:
        "200":
          description: Favorite status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /user/profile:
    get:
      summary: Get user profile
      description: Get the authenticated user's profile
      operationId: getUserProfile
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...